+++ R/utils.R
@@ -1,9 +1,0 @@
-msg <- function(x, startup = FALSE) {
-  if (startup) {
-    if (!isTRUE(getOption("iSEE.quiet"))) {
-      rlang::inform(x, class = "packageStartupMessage")
-    }
-  } else {
-    rlang::inform(x)
-  }
-}